refactored classes, commented booking class
